% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{bplot}
\alias{bplot}
\title{bplot}
\usage{
bplot(data, title = "")
}
\arguments{
\item{data}{input data frame}

\item{title}{plot title}
}
\value{
ggplot object
}
\description{
plot function for biobabble
}
\examples{
data(mooncake)
bplot(mooncake)
}
\author{
Guangchuang Yu
}
