% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DataSet}
\alias{DataSet}
\alias{family}
\alias{lecture}
\alias{mooncake}
\alias{pumpkin}
\alias{apple}
\alias{snowman}
\alias{doctor}
\alias{santa}
\alias{ninja}
\title{Datasets}
\description{
datasets for plotting
}
\keyword{datasets}
